page admin

+++ CONTROLLERS/controller_admin.php
@@ -41,6 +41,10 @@ class controller_admin extends controller_main
     }
     public function Supprimer()
     {
+        if (!empty($_POST['d3']) )
+            throw new Exception('Pseudo vide');
+        if($_POST['d3']->getState()!='admin')
+            throw new Exception('Impossible de supprimer un administrateur');
         $this->user->deleteUser();
     }
 }
